[clean] rename mlpconv3
